fix(test-dx): restore planner ownership and unit import closure
Move repository-relative path decoding into the importable benchmark planner package and keep the GitHub script as a thin compatibility adapter. The planner no longer mutates sys.path to import private CI code.

Also make the unit workflow install the contributor-tool dependency closure that its own unit tests import, rather than depending on ambient pre-commit, YAML, TOML writer, and benchmark packages.
